// src/app/core/utils/operations/admin-op.enum.ts
export enum AdminOpEnum {
  DEPARMENTS = 'DEPARMENTS',
  APPLY_DISCOUNT = 'APPLY_DISCOUNT',
  EMPLZ = 'EMPLZ',
  SYSZ = 'SYSZ',
  WTDZ = 'WTDZ',
  CCSZ = 'CCSZ',
  CHANGE_PRINTER = 'CHANGE_PRINTER',
  CLOSE_BROWSER = 'CLOSE_BROWSER',
  UPDATE_BROWSER = 'UPDATE_BROWSER',
  CANCEL_CHECK = 'CANCEL_CHECK',
  REMOVE_HOLD = 'REMOVE_HOLD',
  REMOVE_TIP = 'REMOVE_TIP',
  AUTH_PENDING = 'AUTH_PENDING',
  BACK_USER = 'BACK_USER',
  SET_USER = 'SET_USER',
  CLOSE_BATCH = 'CLOSE_BATCH',
  CONFIG = 'CONFIG',
  SYSTEM_VERSION = 'SYSTEM_VERSION',
  PREV_SCREEN = 'PREV_SCREEN',
  CHARGE_ACCT_SETUP = 'CHARGE_ACCT_SETUP',
  EMPLOYEE_SETUP = 'EMPLOYEE_SETUP',
  CHANGE_PRICES = 'CHANGE_PRICES',
  EBT_INQUIRY = 'EBT INQUIRY',
  CREDIT_LIMIT = 'Credit Limit',
  REFUND_SALE = 'Refund Sale',
  REVERT_SALE = 'REVERT_SALE',
  WEEKLY_CLOSE = 'WEEKLY_CLOSE',
  CLIENT = 'CLIENT',
  GIFT_CARD = 'GIFT_CARD',
  TIME_WORKED = 'TIME_WORKED',
  UPDATE_PASS = 'UPDATE_PASS',
  UPDATE_CARD = 'UPDATE_CARD',
  DELETE_ACCOUNT = 'DELETE_ACCOUNT',
  UPDATE_NAME = 'UPDATE_NAME',
  UPDATE_POSITION = 'UPDATE_POSITION',
  RANGE_REPORT = "RANGE_REPORT",
  WEEKLY_REPORT = "WEEKLY_REPORT",
  DATER_REPORT = "DATE_RANGE_REPORT",
  CHANGE_COLOR = "CHANGE_COLOR",
  CLOSE_REPORTS = "CLOSE_REPORTS",
  UPDATE_APP = "UPDATE_APP",
  UPDATE = "UPDATE"
}

export enum AdminOpESEnum {
  DEPARMENTS = 'Dept',
  APPLY_DISCOUNT = 'Aplicar Descuento',
  EMPLZ = 'Reporte de Ventas',
  SYSZ = 'Reporte por tipo de pago',
  WTDZ = 'Cierre Diario',
  CCSZ = 'Cierre por Cajero',
  CHANGE_PRINTER = 'Cambiar Impresora',
  UPDATE_BROWSER = 'Actualizar Aplicación',
  CLOSE_BROWSER = 'Cerrar Aplicación',
  CANCEL_CHECK = 'Cancelar Venta',
  REMOVE_HOLD = 'Eliminar venta detenida',
  AUTH_PENDING = 'Autorizaciones Pendientes',
  BACK_USER = 'Back User',
  SET_USER = 'Usuarios',
  CLOSE_BATCH = 'Cierre de Batch',
  CONFIG = 'Configuración',
  SYSTEM_VERSION = 'Versión de la App',
  PREV_SCREEN = 'Prev Screen',
  CHARGE_ACCT_SETUP = 'Acct Charge Setup',
  EMPLOYEE_SETUP = 'Employee Setup',
  CHANGE_PRICES = 'Change Prices',
  EBT_INQUIRY = 'EBT Inquiry',
  CREDIT_LIMIT = 'Credit Limit',
  REFUND_SALE = 'Refund Sale',
  REVERT_SALE = 'Revert Sale',
  WEEKLY_CLOSE = 'Weekly Close',
  CLIENT = 'Client Setup',
  GIFT_CARD = 'Gift Card',
  TIME_WORKED = 'Time Worked',
  UPDATE_PASS = 'Update Password',
  UPDATE_CARD = 'Update Card',
  DELETE_ACCOUNT = 'Eliminar Cuenta',
  UPDATE_NAME = 'Actualizar Nombre',
  RANGE_REPORT = 'Reporte por Rango',
  UPDATE_POSITION = 'Actualizar Posición',
  WEEKLY_REPORT = 'Reporte Semanal',
  CHANGE_COLOR = 'Cambiar Color',
  CLOSE_REPORTS = 'Reporte de Cierre',
  UPDATE_APP = 'Actualizar APP',
  UPDATE = 'Actualizar'
}